project update and calendar add event function

+++ pages/api/appointments/[id].ts
@@ -1,14 +1,28 @@
-import { withApiAuth, supabaseServerClient } from '@supabase/auth-helpers-nextjs';
+import { NextApiHandler } from 'next';
+import { createServerSupabaseClient } from '@supabase/auth-helpers-nextjs';
 
-export default withApiAuth(async function ProtectedRoute(req, res) {
+const ProtectedRoute: NextApiHandler = async (req, res) => {
   const {
     query: { id },
     method,
   } = req;
 
+  // Create authenticated Supabase Client
+  const supabase = createServerSupabaseClient({ req, res });
+  // Check if we have a session
+  const {
+    data: { session },
+  } = await supabase.auth.getSession();
+
+  if (!session)
+    return res.status(401).json({
+      error: 'not_authenticated',
+      description: 'The user does not have an active session or is not authenticated',
+    });
+
   switch (method) {
     case 'DELETE':
-      const { data: deletedData, error: deletedError } = await supabaseServerClient({ req, res })
+      const { data: deletedData, error: deletedError } = await supabase
         .from('appointments')
         .delete()
         .match({ id });
@@ -18,7 +32,7 @@ export default withApiAuth(async function ProtectedRoute(req, res) {
 
       res.json(deletedData);
     case 'PUT':
-      const { data: updatedData, error: updatedError } = await supabaseServerClient({ req, res })
+      const { data: updatedData, error: updatedError } = await supabase
         .from('appointments')
         .update({ ...req.body })
         .match({ id });
@@ -31,4 +45,6 @@ export default withApiAuth(async function ProtectedRoute(req, res) {
       res.setHeader('Allow', ['PUT', 'DELETE']);
       res.status(405).end(`Method ${method} Not Allowed`);
   }
-});
+};
+
+export default ProtectedRoute;
